# backend/vetclinic/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Health Service API",
        "description": "API для системы записи к врачам",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/doctors": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Получить список всех врачей",
                "operationId": "cd066ac893c261ef13773c81ebc8c53e",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Doctor"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Создать нового врача",
                "operationId": "08504526ddafe7c7f25782a3b582fb29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "qualification"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иванов Иван Иванович"
                                    },
                                    "qualification": {
                                        "type": "string",
                                        "example": "Терапевт"
                                    },
                                    "experience": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "https://example.com/avatar.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Врач успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/doctors/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Получить информацию о враче",
                "operationId": "f7d27b7312d2bcfb24da4bf9b7109e21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID врача",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Врач не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Обновить информацию о враче",
                "operationId": "c02c0964a85a8820b7be7c88a00c291a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID врача",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иванов Иван Иванович"
                                    },
                                    "qualification": {
                                        "type": "string",
                                        "example": "Терапевт"
                                    },
                                    "experience": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "https://example.com/avatar.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Врач успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Doctor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Врач не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Удалить врача",
                "operationId": "7b7ef07af466dfd92e8b76108715130f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID врача",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Врач успешно удален"
                    },
                    "404": {
                        "description": "Врач не найден"
                    }
                }
            }
        },
        "/doctors/{doctor}/slots": {
            "get": {
                "tags": [
                    "Doctors",
                    "Slots"
                ],
                "summary": "Получить свободные слоты врача",
                "operationId": "e701dcd9488dddc96dd46c6bd5d4a75f",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "description": "ID врача",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Фильтр по дате начала (от)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Фильтр по дате начала (до)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Slot"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Врач не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Doctor": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иванов Иван Иванович"
                    },
                    "qualification": {
                        "type": "string",
                        "example": "Терапевт"
                    },
                    "experience": {
                        "type": "integer",
                        "example": 10
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Slot": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "doctor_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-25T09:00:00Z"
                    },
                    "end_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-25T09:30:00Z"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "FREE",
                            "BOOKED"
                        ],
                        "example": "FREE"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}